+++ src/main/java/lotto/view/input/InputView.java
@@ -1,26 +1,7 @@
 package lotto.view.input;
 
-import camp.nextstep.edu.missionutils.Console;
-import lotto.domain.lotto.Lotto;
-
 public interface InputView {
-
-    default String readLine(){
-        return Console.readLine();
-    }
-
-    default void printErrorMessage(RuntimeException e) {
-        System.out.println(e.getMessage());
-        System.out.println();
-    }
-
-    default void newLine() {
-        System.out.println();
-    }
-
-    long requestLottoPurchaseAmount();
-
-    Lotto requestWinningLotto();
-
-    int requestBonusLottoNumber(Lotto winningLotto);
+    String readLine();
+    void printErrorMessage(RuntimeException e);
+    void newLine();
 }
