+++ src/main/java/lotto/view/input/ProxyInputView.java
@@ -2,7 +2,7 @@ package lotto.view.input;
 
 import lotto.domain.lotto.Lotto;
 
-public class ProxyInputView implements InputView {
+public class ProxyInputView implements ConsoleInput {
 
     private final LottoGameInputView lottoGameInputView;
 
@@ -10,6 +10,10 @@ public class ProxyInputView implements InputView {
         lottoGameInputView = new LottoGameInputView();
     }
 
+    /**
+     * retry 기능 생각하고 적용해보기
+     * 이 때 retry 넘어서 예외가 나면 똑같이 사용자 입력을 잘못한 예외니까 똑같이 해도될거 같다.
+     */
     @Override
     public long requestLottoPurchaseAmount() {
         try {
